[Consensus Observer] Add simple subscription manager.
